updated product image controller

+++ app/Http/Controllers/ProductImageController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\ProductImage;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProductImageController extends Controller
 {
@@ -14,6 +15,9 @@ class ProductImageController extends Controller
      */
     public function index(Request $request)
     {
+        if (Auth::user()->role === 3) {
+            return redirect()->route('unauthorized');
+        }
         $images = ProductImage::with('file');
         if ($request->input('product_id') !== null) {
             $images = $images->where('product_id', $request->input('product_id'))->get();
@@ -41,7 +45,9 @@ class ProductImageController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request->all());
+        if (Auth::user()->role === 3) {
+            return redirect()->route('unauthorized');
+        }
         $image = ProductImage::create($request->all());
         return response()->json($image);
     }
@@ -54,8 +60,8 @@ class ProductImageController extends Controller
      */
     public function show(ProductImage $productImage)
     {
-        $image = ProductImage::find($productImage->id);
-        return response()->json($image);
+        // $image = ProductImage::find($productImage->id);
+        // return response()->json($image);
     }
 
     /**
@@ -78,9 +84,9 @@ class ProductImageController extends Controller
      */
     public function update(Request $request, ProductImage $productImage)
     {
-        $image = ProductImage::find($productImage->id);
-        $image->update($request->all());
-        return response()->json($image);
+        // $image = ProductImage::find($productImage->id);
+        // $image->update($request->all());
+        // return response()->json($image);
     }
 
     /**
@@ -91,6 +97,9 @@ class ProductImageController extends Controller
      */
     public function destroy(ProductImage $productImage)
     {
+        if (Auth::user()->role === 3) {
+            return redirect()->route('unauthorized');
+        }
         return ProductImage::destroy($productImage->id);
     }
 }
